Ensure testing utilities are included in the built/installed package
Move Flow's testing utilities from 'resolwe.flow.tests' to
'resolwe.flow.utils.test' and from 'resolwe.permissions.tests.base' to
'resolwe.permissions.utils.test'.
Properly include documentation for testing utilities and data purge in
the Reference documentation.
