+++ docs/CHANGELOG.rst
@@ -49,6 +49,9 @@ Changed
 - Put packaging tests in a separate Tox testing environment
 - Put linters (pylint, pep8) into a separate Tox testing environment
 - Drop django-jenkins package since we no longer use Jenkins for CI
+- Move testing utilities from ``resolwe.flow.tests`` to
+  ``resolwe.flow.utils.test`` and from ``resolwe.permissions.tests.base`` to
+  ``resolwe.permissions.utils.test``
 
 Fixed
 -----
