+++ docs/ref.rst
@@ -3,4 +3,5 @@ Reference
 =========
 
 .. automodule:: resolwe.permissions.shortcuts
-.. automodule:: resolwe.permissions.tests
+.. automodule:: resolwe.permissions.utils
+.. automodule:: resolwe.flow.utils
