--- resolwe/permissions/tests/__init__.py
@@ -1,8 +1,0 @@
-"""
-=======
-Testing
-=======
-
-.. automodule:: resolwe.permissions.tests.base
-
-"""
