+++ resolwe/permissions/utils/__init__.py
@@ -1,0 +1,10 @@
+"""
+=====================
+Permissions Utilities
+=====================
+
+.. automodule:: resolwe.permissions.utils.test
+   :members:
+
+"""
+from __future__ import absolute_import, division, print_function, unicode_literals
