+++ resolwe/permissions/utils/test.py
@@ -1,5 +1,7 @@
 """
-.. autoclass:: ResolweAPITestCase
+=======
+Testing
+=======
 
 """
